Switch to considering multiple possible series for energy and OPEX and return multiple series for LCOE
Now the opex_bom and the energy_record just need the project_year column. Every other column is considered to be a cost or energy, respectively.

The outputs are now arrays, one for each column found in the opex_bom or the energy_record. If both are given they are assumed to have the same number of columns.